requirements.txt
livekit-agents
livekit-plugins-assemblyai
livekit-plugins-openai
livekit-plugins-google
livekit-plugins-silero
python-dotenv
pydantic-ai-slim[openai,mcp]
firecrawl-py